Extract getPRsToPaintPerSHAs
